added:.gitignore , constants.js , index.js , helpers.js ,modified:package.json, diamond.js , equilateral.js , permissions.js , validation.js
