+++ extra/pattern/index.js
@@ -1,0 +1,3 @@
+import diamond from './diamond'
+import equilateral from './equilateral'
+export{diamond,equilateral};
